initial commit
This is an initial PoC that launches VLC with the appropriate
permissions. As of now, win32 APIs aren't aware of how to access files
properly even when proper security_capabilities are applied on them, so
you have to add an ACL manually to whatever files you want to give perms
to with `icacls.exe`. The short-term goal is adding support to
automatically add and remove DACL's to access these files, and let them
be deleted when not in use.

Ideally, we need to find a way to make sure DACLs don't stick around in
the event of a crash. A hacky way would be to just keep a log of
modified DACLs somewhere and undo them the next time we run (for the
event of a crash), and also to try to delete them when we exit. A good
solution would be to figure out if we can create an ephermal DCL that
only exists in memory while a process lives through some API in win32.